pyproject.toml
[project]
name = "drover"
version = "0.1.0"
description = "Document classification CLI that classifies files into organized folder structures"
readme = "README.md"
requires-python = ">=3.11.3"
license = {text = "MIT"}
dependencies = [
    "langchain>=0.3",
    "langchain-community>=0.3",
    "langchain-openai>=0.2",
    "langchain-ollama>=0.2",
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "bandit>=1.7.5",
]

[project.scripts]
drover = "drover.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/drover"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
